% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_flag_list}
\alias{make_flag_list}
\title{Create list of conditional flags}
\usage{
make_flag_list(tokenlist)
}
\arguments{
\item{tokenlist, }{named list with at least 1 item}
}
\value{
list, with one item per \code{tokenlist} item
}
\description{
This function identifies whether an item in a named list is \code{NA} or not,
and creates a named list with respective \code{TRUE} / \code{FALSE} values.
This list can be appended to an existing whisker tokens list, and used in templates for conditional
generation of text snippets and code chunks.
The names in the generated list are of the form \code{previousname_roar_flag}
}
