% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_timestamp}
\alias{make_timestamp}
\title{Convenience function to create a universal time stamp}
\usage{
make_timestamp(x, timestamp_cols)
}
\arguments{
\item{x}{data frame which contains \code{timestamp_cols}}

\item{timestamp_cols}{character, column(s) defining time stamp and must be in \code{x}}
}
\value{
x, with additional column \code{timestamp_roar}
}
\description{
Grab one or more time-defining columns and create a universal time stamp,
in character format, ideally as "YYYY-mm-dd HH:MM:SS)
This stamp can be re-read in subsequent steps, and the time zone adjusted as necessary.
}
