% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_meta_strings}
\alias{parse_meta_strings}
\title{Split up strings in meta table}
\usage{
parse_meta_strings(string)
}
\arguments{
\item{string}{}
}
\value{
character, (split) \code{string}
}
\description{
For multi-column data adjustments/details, e.g. definition of time stamp columns, a meta string
can be provided in a single column of the meta table (e.g. \code{time_stamp_columns}), preventing the
meta table from cluttering.
This is can be leveraged for in templates to construct objects from meta strings.
A meta string can be set to either \code{NA}, a single value (\code{"string"}) or a multi-column string
(\code{"string1/string2/string3"}).
This function splits these strings, and allows e.g. pasting multiple columns together.
}
