% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readskip.csv}
\alias{readskip.csv}
\title{Skip read}
\usage{
readskip.csv(path, ..., skip_internal = NA)
}
\arguments{
\item{...}{additional arguments passed to \code{read.csv()}.}

\item{skip_internal}{numeric, vector of rows to skip \strong{below} headers. These row numbers correspond to
the raw input, i.e. if headers are in row one, and additional information in row 2, \code{skip_internal = 2}.}
}
\value{
data.frame read from \code{path}
}
\description{
This function reads a \code{.csv} while skipping specified rows (e.g. meta table included beneath headers).
Note, that \code{stringsAsFactors = FALSE} is assumed.
}
