\name{RoarDataset-class.Rd}
\alias{RoarDataset-class}
\docType{data}
\title{
NON EMPTY TITLE

}
\description{

}
\usage{data(RoarDataset-class.Rd)}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{

}
\examples{

}
\keyword{datasets}
