% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_meta}
\alias{update_meta}
\title{Wrapper for updating a meta table}
\usage{
update_meta(meta_table_path, idx, update_list, ...)
}
\arguments{
\item{meta_table_path}{character, path to meta table (assumed relative to project dir, but can be absolute)}

\item{idx}{numeric, length 1, giving row index of the meta table corresponding to dataset being processed}

\item{update_list}{named list, length 1, where name equals the target column in meta table, and value is the adjustment}

\item{...}{reserved for including automated git commits to track changes to meta table}
}
\value{
nothing
}
\description{
Wrapper for updating a meta table
}
