% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_template}
\alias{view_template}
\title{Print internal template to console}
\usage{
view_template(template_name = "prep_script.R")
}
\arguments{
\item{template_name}{character, name of internal tempalte (currently only \code{prep_script.R} )}
}
\value{
all available internal templates
}
\description{
Print internal template to console
}
