\name{roar-package}
\alias{roar-package}
\alias{roar}
\docType{package}
\title{
Identify differential APA usage from RNA-seq alignments
}
\description{
Identify preferential usage of APA sites, 
comparing two biological conditions, starting from known alternative sites and
alignments obtained from standard RNA-seq experiments.
}
\details{
\tabular{ll}{
Package: \tab roar\cr
Type: \tab Package\cr
Version: \tab 0.1.1 \cr
Date: \tab 2013-05-04\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Elena Grassi <grassi.e@gmail.com>
}
\references{
 TODO
}
\keyword{ package }

